+++ code/ts/library/src/plugins/official/attributes/core/mergeSignals.ts
@@ -1,16 +1,16 @@
 // Authors: Delaney Gillilan
 // Icon: material-symbols:home-storage
-// Slug: Store signals into a singleton per page
-// Description: This action stores signals into a singleton per page. This is useful for storing signals that are used across multiple components.
+// Slug: Merge signals into a singleton per page
+// Description: This action signalss signals into a singleton per page. This is useful for storing signals that are used across multiple components.
 
 import {
     AttributeContext,
     AttributePlugin,
     RegexpGroups,
 } from "../../../../engine";
-import { storeFromPossibleContents } from "../../../../utils/signals";
+import { signalsFromPossibleContents } from "../../../../utils/signals";
 
-// Setup the global store
+// Merge into singleton signals
 export const MergeSignals: AttributePlugin = {
     pluginType: "attribute",
     name: "mergeSignals",
@@ -19,11 +19,11 @@ export const MergeSignals: AttributePlugin = {
         pre: [
             {
                 pluginType: "preprocessor",
-                name: "store",
+                name: "signals",
                 regexp: /(?<whole>.+)/g,
                 replacer: (groups: RegexpGroups) => {
                     const { whole } = groups;
-                    return `Object.assign({...ctx.store()}, ${whole})`;
+                    return `Object.assign({...ctx.signals()}, ${whole})`;
                 },
             },
         ],
@@ -31,8 +31,8 @@ export const MergeSignals: AttributePlugin = {
     allowedModifiers: new Set(["ifmissing"]),
     onLoad: (ctx: AttributeContext) => {
         const possibleMergeSignals = ctx.expressionFn(ctx);
-        const actualMergeSignals = storeFromPossibleContents(
-            ctx.store(),
+        const actualMergeSignals = signalsFromPossibleContents(
+            ctx.signals(),
             possibleMergeSignals,
             ctx.modifiers.has("ifmissing"),
         );