+++ CHANGELOG.md
@@ -4,5 +4,5 @@
 
 ### Changed
 
-- Renamed the `data-store` attribute to `data-merge-signals`.
+- Renamed the `data-signals` attribute to `data-merge-signals`.
 - Renamed the `data-model` attribute to `data-bind`.